Rename ExposedSpec to ExposedComponent.
