+++ lib/exposedcomponent.js
