+++ lib/injectedcontainer.js
@@ -1,6 +1,6 @@
 // Load modules.
 var path = require('canonical-path')
-  , ExposedSpec = require('./exposedspec');
+  , ExposedComponent = require('./exposedcomponent');
 
 
 /**
@@ -73,10 +73,10 @@ InjectedContainer.prototype.specs = function(q, ins) {
     rid = path.relative(this._ns, spec.id);
     if (ins === true && rid.indexOf('../') == 0) { continue; }
     if (!q) {
-      exposed.push(new ExposedSpec(spec));
+      exposed.push(new ExposedComponent(spec));
     } else if (spec.implements.some(function(i) { return i == q })) {
       // The spec implements one of the requested interfaces.
-      exposed.push(new ExposedSpec(spec));
+      exposed.push(new ExposedComponent(spec));
     }
   }
   return exposed;
